pyproject.toml
[project]
name = "fake-npc-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
]
